requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
controlnet-aux>=0.0.7
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
ultralytics>=8.0.0
git+https://github.com/openai/CLIP.git
realesrgan>=0.3.0
basicsr>=1.4.2
